Refactor : Fragment에서 UI와 관련되지 않은 코드들은 ViewModel로 옮기는 중
